vpp-plugin: implement vlib::BufferRef::vnet_feature_next in Rust
Although in theory cross-language LTO is possible, it seems that it's
sensitive to the compilers used (cannot mix and match gcc for C and
LLVM backend for Rust, for example) and even the major versions of the
toolchains (e.g. clang-20 may not be compatible with an LLVM 18.0.0
backend used by Rust, although currently LLVM 21.1.2 is used by the
most recent stable rustc release).

For these reasons, and the fact that implementing
vlib::BufferRef::vnet_feature_next itself is trivial (even if it
depends on further type abstractions), implement it natively in Rust
so it can be inlined into node functions, as it will be commonly used
in nodes invoked from feature arcs.

Move the implementation to vpp-plugin/src/vnet/buffer.rs to better
match where it's implement in the VPP source code.

This improves the performance of the example node no-match case from
1.80e1 clocks per packet to 1.39e1.