node_generic: tweaks
Prefer FeatureNextNode::NextFeature over
FeatureNextNode::DefinedNode(next) since the NextFeature is more
likely to be the common case. Note that this doesn't really affect the
instructions that are generated, but it does make the code match the
generated instructions for at least the example node.

In addition, don't store the buffer length in a variable for returning
from the generic node function, but instead get it again from the
frame. The compiler isn't free to optimise this because it doesn't
know whether the frame C pointer is aliased from other pointers used
in the function. It's not obvious whether this is an improvement in
performance as previously b_len would need to be stored in a register
(or on the stack) whereas now frame needs to be stored in a register
(or on the stack), but retrieving again from the frame at least
matches what C node implementations do and so removes one potential
for performance difference when comparing.