buffer: optimise vlib::MainRef::get_buffers performance
Although the compiler already does vectorise
vlib::MainRef::get_buffers for architectures with AVX-512F
instructions, it only does so using 4 x u32x8 strides with the last 32
elements being done with non-vectorised instructions.

We can do better by strongly hinting to the compiler to use 64 element
strides and by changing the preconditions of the function such that
vectorised instructions are used, even for the last 8 elements. In
order to avoid bizarre instructions emitted by LLVM for the last 64
elements then the loop is manually unrolled for the n >= 32 and n >=
16 cases. These instructions don't seem to have any performance
impact, but hinder reading the disassembly, so are worth avoiding.

This improves the performance of the example node no-match case from
1.25e1 clocks per packet to ~1.19e1.