node_generic: avoid a bounds check
The compiler adds in a check to ensure that when taking
&nexts[..b.len()] that b.len() <= nexts.len(), and panicking
otherwise. However, this is in the fast path so we don't want such a
check. Therefore, optimise by using slice::from_raw_parts() to remove
the bounds check, which is safe because the function already controls
the size of both nexts and b.

This improves the performance of the example node no-match case from
1.39e1 clocks per packet to 1.34e1.