vpp-plugin-api-gen: support stream messages
Add support for stream messages in message handler functions by taking
in a Stream<T> parameter where T is the type of the stream
message. This allows handlers to easily send messages and provides
type consistency with the .api file definition.

Add tests for stream messages and messages with interesting replies.