Add support for variable-length arrays in messages
Add support for variable-length arrays in messages by adding accessors
into the variable-length arrays, using Unaligned* types for primitives
(other than u8).

Since now the count field is made use of when creating a slice into
the VLA, and this since this needs to be consistent with the memory
backing the object, both the accessors and now the endian swap which
accesses this data need to be made unsafe, which represents a
backwards-incompatible change for both vpp-plugin and
vpp-plugin-api-gen (due to the impact on integrating with the
generated code).

In addition, Stream::send_message() is now unsafe because it may be
used to send a message containing a VLA and access the VLA both
directly for the endian swap and indirectly via VPP code in actually
sending the message.

Note that types containing VLAs are still not supported at this point,
only directly in messages.

Also note that it would be nice if unsafe functions didn't have to be
used to access the variable-length array, but that can't be done
inside the type system until the ptr_metadata feature
(https://github.com/rust-lang/rust/issues/81513) is stabilised.