vpp-plugin-api-gen: generate API enums as packed
Rather than allowing API enum types to have natural alignment for
their underlying primitive type, generate API enum types with
repr(packed) so that they have 1-byte alignment like other types
generated from API files. This has the benefit that endian_swap (and
calc_size in the future) methods/functions are simpler, plus the types
become simpler to use when embedded in other API types/messages.

The only downside is that on certain architectures and outside of use
in API types/messages then less efficient code may be generated by the
Rust compiler, but this is deemed to be a corner case and not worth
optimising for.