vpp-plugin-api-gen: make f64 endian swap explicitly no-op
The current code accidentally makes endian swap a no-op by converting
from native to big-endian bytes and then back again. This wasn't
caught by tests because it turns out that it's a no-op in VPP for not
entirely clear reasons (floats do have endianness which is defined by
the platform's ISA, even if it isn't defined by IEEE).

In any case, we need to be compatible with what VPP and the Python
vpp_papi assume, so make the fact that it's a no-op explicit. This has
the benefit of simplifying the code for endian-swapping f64, and
avoiding the need to add code to handle endian swaps for f64 VLAs in
the future.