vlibapi: allow constructing dynamically-sized messages
In order to support variable-length arrays at the end of messages,
which cannot be sized statically and therefore cannot use normal
types, add a Message::new_bytes() method that allocates a message of
the given length and returns a Message<u8> object. This can then be
transposed into a suitable type in the future by API code.

Whilst there, also don't force only use of Sized objects with Message,
even though at the moment there is no use for this.

In addition, expose additional traits on Message objects to make them
easier to use, particularly the format ones for use in debugging.