vlibapi: add Unaligned* types
The u16, i16, u32, i32, u64, i64 and f64 types types in Rust have an
alignment of > 1 byte, whereas the API messages (and their fields) are
unaligned. This isn't too much a problem for fields, but is a problem
for representing variable-length arrays at the end of messages as a
slice produced from an accessor, since it must be built from a
suitably aligned pointer.

So add Unaligned* types for these primitive types to allow returning
slices for variable-length arrays.